requirements.txt
sreamlit
PyPDF2
google.generativeai
python-dotenv
